requirements.txt
transformers
datasets
torch
mlflow
fastapi>=0.100.0
uvicorn
gradio
pandera
pydantic>=2.0
slowapi
prometheus-fastapi-instrumentator
python-json-logger
bandit
pip-audit
detect-secrets
scikit-learn
pandas
numpy
pyyaml
psutil
httpx
pytest-asyncio
